requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pandas==2.1.1
numpy==1.26.0
requests==2.31.0
google-generativeai==0.3.0
langchain==0.1.0
pydantic==2.4.2
aiohttp==3.9.0
streamlit==1.28.1
